+++ packages/content-source-maps/vite.config.ts
@@ -1,4 +1,6 @@
+import { resolve } from 'path';
 import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
 import { configDefaults } from 'vitest/config';
 
 export default defineConfig({
@@ -6,4 +8,18 @@ export default defineConfig({
     environment: 'jsdom',
     exclude: [...configDefaults.exclude],
   },
+  build: {
+    sourcemap: true,
+    outDir: './dist',
+    lib: {
+      entry: resolve(__dirname, 'src/index.ts'),
+      fileName: 'index',
+      formats: ['cjs', 'es'],
+    },
+  },
+  plugins: [
+    dts({
+      exclude: ['**/__tests__/**'],
+    }),
+  ],
 });